Complete ParRepacker

+++ ParRepacker/ParRepacker.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Linq;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -18,18 +19,31 @@ namespace ParRepacker
         {
             var parTasks = new List<Task>();
 
+            string pathToParlessMods = Path.Combine(GamePath.GetModsPath(), "Parless");
+
+            if (Directory.Exists(pathToParlessMods))
+            {
+                // Clean up the /mods/Parless/ directory before processing the dictionary
+                // TODO: If .parmeta checks are implemented, this should be changed
+                Directory.Delete(pathToParlessMods, true);
+            }
+
             foreach (KeyValuePair<string, List<string>> parModPair in parDictionary)
             {
-                parTasks.Add(Task.Run(() => RepackPar(parModPair.Key, parModPair.Value)));
+                RepackPar(parModPair.Key, parModPair.Value);
+                //parTasks.Add(Task.Run(() => RepackPar(parModPair.Key, parModPair.Value)));
             }
 
             await Task.WhenAll(parTasks).ConfigureAwait(false);
         }
 
-        private static async Task RepackPar(string par, List<string> mods)
+        //private static async Task RepackPar(string par, List<string> mods)
+        private static void RepackPar(string parName, List<string> mods)
         {
-            string pathToPar = Path.Combine(GamePath.GetDataPath(), par + ".par");
-            string pathToModPar = Path.Combine(GamePath.GetModsPath(), "Parless", par + ".par");
+            parName = parName.TrimStart(Path.DirectorySeparatorChar);
+
+            string pathToPar = Path.Combine(GamePath.GetDataPath(), parName + ".par");
+            string pathToModPar = Path.Combine(GamePath.GetModsPath(), "Parless", parName + ".par");
             string pathToModParMeta = pathToModPar + "meta";
 
             // Dictionary of fileInPar, ModName
@@ -38,7 +52,7 @@ namespace ParRepacker
             // Populate fileDict with the files inside each mod
             foreach (string mod in mods)
             {
-                foreach (string modFile in GetModFiles(par, mod))
+                foreach (string modFile in GetModFiles(parName, mod))
                 {
                     fileDict.TryAdd(modFile, mod);
                 }
@@ -102,8 +116,22 @@ namespace ParRepacker
             {
                 string pathToTempPar = pathToModPar + "temp";
 
-                // Make sure that the .partemp directory is empty
-                Directory.Delete(pathToTempPar, true);
+                if (File.Exists(pathToModPar))
+                {
+                    // Make sure that the .partemp directory is empty
+                    File.Delete(pathToModPar);
+                }
+
+                if (Directory.Exists(pathToTempPar))
+                {
+                    // Make sure that the .partemp directory is empty
+                    Directory.Delete(pathToTempPar, true);
+                    Directory.CreateDirectory(pathToTempPar);
+                }
+                else
+                {
+                    Directory.CreateDirectory(pathToTempPar);
+                }
 
                 // Copy each file in the mods to the .partemp directory
                 foreach (KeyValuePair<string, string> fileModPair in fileDict)
@@ -111,35 +139,88 @@ namespace ParRepacker
                     if (fileModPair.Value.StartsWith("Parless_"))
                     {
                         File.Copy(
-                            Path.Combine(GamePath.GetDataPath(), par.Insert(int.Parse(fileModPair.Value.Substring(8)), ".parless"), fileModPair.Key),
+                            Path.Combine(GamePath.GetDataPath(), parName.Insert(int.Parse(fileModPair.Value.Substring(8)), ".parless"), fileModPair.Key),
                             Path.Combine(pathToTempPar, fileModPair.Key),
                             true);
                     }
                     else
                     {
                         File.Copy(
-                            GamePath.GetModPathFromDataPath(fileModPair.Value, Path.Combine(par, fileModPair.Key)),
+                            GamePath.GetModPathFromDataPath(fileModPair.Value, Path.Combine(parName, fileModPair.Key)),
                             Path.Combine(pathToTempPar, fileModPair.Key),
                             true);
                     }
                 }
 
-                // Create a node from the .partemp directory and write the par to pathToModPar
-                string nodeName = new DirectoryInfo(pathToTempPar).Name;
-                Node node = ReadDirectory(pathToTempPar, nodeName);
+                ParArchiveReaderParameters readerParameters = new ParArchiveReaderParameters
+                {
+                    Recursive = false,
+                };
 
-                node.SortChildren((x, y) => string.CompareOrdinal(x.Name.ToLowerInvariant(), y.Name.ToLowerInvariant()));
-
-                Directory.CreateDirectory(Path.GetDirectoryName(Path.GetFullPath(pathToModPar)));
-                node.TransformWith<ParArchiveWriter, ParArchiveWriterParameters>(new ParArchiveWriterParameters {
+                ParArchiveWriterParameters writerParameters = new ParArchiveWriterParameters
+                {
                     CompressorVersion = 0,
                     OutputPath = pathToModPar,
                     IncludeDots = true,
-                });
 
-                // TODO: Write .parmeta file here
+                };
+
+                // Create a node from the .partemp directory and write the par to pathToModPar
+                Console.Write("Reading input directory... ");
+                string nodeName = new DirectoryInfo(pathToTempPar).Name;
+                Node node = ReadDirectory(pathToTempPar, nodeName);
+                Console.WriteLine("DONE!");
+
+                bool overwrite = false;
+
+                // Check if overwrite.txt exists
+                for (int i = 0; i < node.Children.Count; i++)
+                {
+                    if (node.Children[i].Name == "overwrite.txt")
+                    {
+                        node.Children[i].Dispose();
+                        node.Remove(node.Children[i]);
+
+                        overwrite = true;
+                        break;
+                    }
+                }
+
+                // Store a reference to the nodes in the container to dispose of them later, as they are not disposed properly
+                NodeContainerFormat containerNode = node.GetFormatAs<NodeContainerFormat>();
+
+                if (overwrite)
+                {
+                    node.SortChildren((x, y) => string.CompareOrdinal(x.Name.ToLowerInvariant(), y.Name.ToLowerInvariant()));
+                    node.TransformWith<ParArchiveWriter, ParArchiveWriterParameters>(writerParameters);
+                }
+                else
+                {
+                    Console.Write("Reading PAR file... ");
+                    Node par = NodeFactory.FromFile(pathToPar, Yarhl.IO.FileOpenMode.Read);
+                    par.TransformWith<ParArchiveReader, ParArchiveReaderParameters>(readerParameters);
+                    writerParameters.IncludeDots = par.Children[0].Name == ".";
+                    Console.WriteLine("DONE!");
+
+                    Console.Write("Adding files... ");
+                    node.GetFormatAs<NodeContainerFormat>().MoveChildrenTo(writerParameters.IncludeDots ? par.Children[0] : par, true);
+                    par.SortChildren((x, y) => string.CompareOrdinal(x.Name.ToLowerInvariant(), y.Name.ToLowerInvariant()));
+                    Console.WriteLine("DONE!");
+
+                    writerParameters.IncludeDots = false;
+
+                    Console.WriteLine("Creating PAR (this may take a while)... ");
+                    Directory.CreateDirectory(Path.GetDirectoryName(Path.GetFullPath(pathToModPar)));
+                    par.TransformWith<ParArchiveWriter, ParArchiveWriterParameters>(writerParameters);
+                    par.Dispose();
+                }
+
+                // TODO: Write .parmeta file here, using the containerNode
 
                 node.Dispose();
+                containerNode.Root.Dispose();
+
+                Console.WriteLine("DONE!");
 
                 // Remove the .partemp directory
                 Directory.Delete(pathToTempPar, true);
@@ -148,13 +229,17 @@ namespace ParRepacker
 
         private static List<string> GetModFiles(string par, string mod)
         {
+            List<string> result;
             if (mod.StartsWith("Parless_"))
             {
                 // Get index of ".parless" in par path
-                return GetModFiles(Path.Combine(GamePath.GetDataPath(), par.Insert(int.Parse(mod.Substring(8)), ".parless")));
+                result = GetModFiles(Path.Combine(GamePath.GetDataPath(), par.Insert(int.Parse(mod.Substring(8)), ".parless")));
             }
 
-            return GetModFiles(GamePath.GetModPathFromDataPath(mod, par));
+            result = GetModFiles(GamePath.GetModPathFromDataPath(mod, par));
+
+            // Get file path relative to par
+            return result.Select(f => f.Substring(f.IndexOf(par) + par.Length + 1)).ToList();
         }
 
         private static List<string> GetModFiles(string path)
@@ -162,7 +247,7 @@ namespace ParRepacker
             List<string> files = new List<string>();
 
             // Add files in current directory
-            files.AddRange(Directory.GetFiles(path));
+            files.AddRange(Directory.GetFiles(path).Select(f => GamePath.GetDataPathFrom(f)));
 
             // Get files for all subdirectories
             foreach (string folder in Directory.GetDirectories(path))
